improve mock message sending
