+++ FlowCrypt/Functionality/Api/Remote Private Key Apis/EmailKeyManagerApi.swift
@@ -83,8 +83,7 @@ actor EmailKeyManagerApi: EmailKeyManagerApiType {
         for privateKey in decryptedPrivateKeysResponse.privateKeys {
             let parsedPrivateKey = try await core.parseKeys(armoredOrBinary: privateKey.decryptedPrivateKey.data())
             // todo - check that parsedPrivateKeys don't contain public keys
-            let isKeyDecrypted = parsedPrivateKey.keyDetails
-                .compactMap { $0.isFullyDecrypted }
+            let isKeyDecrypted = parsedPrivateKey.keyDetails.compactMap(\.isFullyDecrypted)
             if isKeyDecrypted.contains(false) {
                 throw EmailKeyManagerApiError.keysAreNotDecrypted
             }
