+++ FlowCrypt/Functionality/Api/Remote Pub Key Apis/WkdApi.swift
@@ -6,7 +6,6 @@
 //  Copyright © 2017-present FlowCrypt a. s. All rights reserved.
 //
 
-import CryptoKit
 import FlowCryptCommon
 
 protocol WkdApiType {
