+++ appium/api-mocks/apis/attester/attester-endpoints.ts
@@ -8,7 +8,7 @@ import { AttesterConfig, MockConfig } from '../../lib/configuration-types';
 
 export class AttesterErr extends HttpErr {
   public formatted = (): unknown => {
-    return { // follows Attester errror format
+    return { // follows Attester error format
       error: {
         "code": this.statusCode,
         "message": this.message,
@@ -263,5 +263,20 @@ FXJldm9rZWRAZmxvd2NyeXB0LmNvbcJeBBMWCgAGBQJhbwFrAAoJEF8Vl4kQ
 oXgKecoBALdrD8nkptLlT8Dg4cF+3swfY1urlbdEfEvIjN60HRDLAP4w3qeS
 zZ+OyuqPFaw7dM2KOu4++WigtbxRpDhpQ9U8BQ==
 =bMwq
+-----END PGP PUBLIC KEY BLOCK-----`,
+  dmitry: `-----BEGIN PGP PUBLIC KEY BLOCK-----
+Version: FlowCrypt iOS 0.2 Gmail Encryption
+Comment: Seamlessly send and receive encrypted email
+
+xjMEYrr+kxYJKwYBBAHaRw8BAQdAl3dSHtMwIEq7pqZ68AZqUKsCElLBET6K
+JwTh8v6bfXvNKkRtaXRyeSBhdCBGbG93Q3J5cHQgPGRtaXRyeUBmbG93Y3J5
+cHQuY29tPsJ3BBAWCgAfBQJiuv6TBgsJBwgDAgQVCAoCAxYCAQIZAQIbAwIe
+AQAKCRCVBUkqsjSkNv6dAP0RAf4+MmLIeFcB1dqwKgj/5cIZhdlt8BARvNLD
+kgk1hgD+O4t3VtYG3aK7VOdc8RYEVzCIBWGA0xF7M2o7zp9C0wPOOARiuv6T
+EgorBgEEAZdVAQUBAQdAiyNEwp+NGJNnDU/LBlHNh+lZbTIGlDd050EceDkL
+13YDAQgHwmEEGBYIAAkFAmK6/pMCGwwACgkQlQVJKrI0pDbAGwD/eLXL8I+y
+udeNtOHMPfrAvfOm1fPYWjRr58M64Osa1RAA/RySIH6WSBoyw79fQfI+uklB
+xGiZ6dc7dKcqxEEVIJMM
+=njJq
 -----END PGP PUBLIC KEY BLOCK-----`
 };
