+++ appium/api-mocks/apis/google/exported-messages/message-export-17d7a76924f539ec.json
@@ -1,0 +1,54 @@
+{
+  "acctEmail": "e2e.enterprise.test@flowcrypt.com",
+  "full": {
+    "id": "17d7a76924f539ec",
+    "threadId": "17d7a76924f539ec",
+    "labelIds": [
+      "INBOX"
+    ],
+    "snippet": "On reply, it should show reply@domain.com not flowcrypt.compatibility@gmail.com",
+    "payload": {
+      "partId": "",
+      "mimeType": "text/plain",
+      "filename": "",
+      "headers": [
+        {
+          "name": "Subject",
+          "value": "Honor reply-to address - plain"
+        },
+        {
+          "name": "From",
+          "value": "flowcrypt.compatibility@gmail.com"
+        },
+        {
+          "name": "To",
+          "value": "e2e.enterprise.test@flowcrypt.com"
+        },
+        {
+          "name": "Reply-To",
+          "value": "reply@domain.com"
+        }
+      ],
+      "body": {
+        "size": 80,
+        "data": "T24gcmVwbHksIGl0IHNob3VsZCBzaG93IHJlcGx5QGRvbWFpbi5jb20Kbm90IGZsb3djcnlwdC5jb21wYXRpYmlsaXR5QGdtYWlsLmNvbQo="
+      }
+    },
+    "sizeEstimate": 226,
+    "historyId": "170894",
+    "internalDate": "1638437131168"
+  },
+  "attachments": {},
+  "raw": {
+    "id": "17d7a76924f539ec",
+    "threadId": "17d7a76924f539ec",
+    "labelIds": [
+      "INBOX"
+    ],
+    "snippet": "On reply, it should show reply@domain.com not e2e.enterprise.test@flowcrypt.com",
+    "sizeEstimate": 226,
+    "raw": "U3ViamVjdDogSG9ub3IgcmVwbHktdG8gYWRkcmVzcyAtIHBsYWluCkZyb206IGZsb3djcnlwdC5jb21wYXRpYmlsaXR5QGdtYWlsLmNvbQpUbzogZmxvd2NyeXB0LmNvbXBhdGliaWxpdHlAZ21haWwuY29tClJlcGx5LVRvOiByZXBseUBkb21haW4uY29tCgpPbiByZXBseSwgaXQgc2hvdWxkIHNob3cgcmVwbHlAZG9tYWluLmNvbQpub3QgZmxvd2NyeXB0LmNvbXBhdGliaWxpdHlAZ21haWwuY29tCg==",
+    "historyId": "170894",
+    "internalDate": "1638437131168"
+  }
+}
