+++ appium/api-mocks/apis/google/google-messages.ts
@@ -1,4 +1,4 @@
 /* ©️ 2016 - present FlowCrypt a.s. Limitations apply. Contact human@flowcrypt.com */
 
 export type GoogleMockAccountEmail = 'e2e.enterprise.test@flowcrypt.com' | 'test2@example.net';
-export type GoogleMockMessage = 'CC and BCC test' | 'Test 1' | 'Signed and encrypted message';
+export type GoogleMockMessage = 'CC and BCC test' | 'Test 1' | 'Signed and encrypted message' | 'Honor reply-to address - plain' | 'email with text attachment' | 'Message with cc and multiple recipients and attachment';
