+++ appium/api-mocks/mock.ts
@@ -82,7 +82,7 @@ export class MockApi {
       }
     };
     mockApi.ekmConfig = {
-      returnKeys: [ekmKeySamples.key0.prv, ekmKeySamples.e2eValidKey.prv, ekmKeySamples.key1.prv]
+      returnKeys: [ekmKeySamples.e2eMock.prv] //[ekmKeySamples.e2eValidKey.prv] // [ekmKeySamples.key0.prv, ekmKeySamples.e2eValidKey.prv, ekmKeySamples.key1.prv]
     }
     mockApi.googleConfig = {
       accounts: {
@@ -91,7 +91,7 @@ export class MockApi {
             { displayName: 'Dima Flowcrypt', email: 'dmitry@flowcrypt.com' },
             { displayName: 'Demo', email: 'demo@flowcrypt.com' }
           ],
-          messages: ['CC and BCC test', 'Test 1', 'Signed and encrypted message'],
+          messages: ['CC and BCC test', 'Test 1', 'Signed and encrypted message', 'Honor reply-to address - plain', 'email with text attachment', 'Message with cc and multiple recipients and attachment'],
         }
       }
     }
@@ -101,7 +101,9 @@ export class MockApi {
         'revoked@flowcrypt.com': attesterPublicKeySamples.revoked,
         'robot@flowcrypt.com': attesterPublicKeySamples.valid,
         'test2@example.net': attesterPublicKeySamples.valid,
-        'dmitry@flowcrypt.com': attesterPublicKeySamples.valid
+        'dmitry@flowcrypt.com': attesterPublicKeySamples.dmitry,
+        'valid@domain.test': attesterPublicKeySamples.valid,
+        'e2e.enterprise.test@flowcrypt.com': ekmKeySamples.e2eMock.pub!
       }
     };
     return mockApi
