--- appium/tests/specs/live/inbox/ReadAttachmentEmail.spec.ts
@@ -1,76 +1,0 @@
-import {
-  SplashScreen,
-  SetupKeyScreen,
-  MailFolderScreen,
-  EmailScreen,
-  AttachmentScreen,
-  SearchScreen
-} from '../../../screenobjects/all-screens';
-
-import { CommonData } from '../../../data';
-
-describe('INBOX: ', () => {
-
-  it('user is able to view encrypted email with attachment', async () => {
-
-    const senderName = CommonData.sender.name;
-    const emailSubject = CommonData.encryptedEmailWithAttachment.subject;
-    const emailText = CommonData.encryptedEmailWithAttachment.message;
-    const attachmentName = CommonData.encryptedEmailWithAttachment.attachmentName;
-    const attachmentNameWithoutExtension = attachmentName.substring(0, attachmentName.lastIndexOf('.')) || attachmentName;
-    const encryptedAttachmentName = CommonData.encryptedEmailWithAttachment.encryptedAttachmentName;
-
-    const wrongPassPhrase = 'wrong';
-    const correctPassPhrase = CommonData.account.passPhrase;
-    const bundleId = CommonData.bundleId.id;
-
-    await SplashScreen.login();
-    await SetupKeyScreen.setPassPhrase();
-    await MailFolderScreen.checkInboxScreen();
-
-    await MailFolderScreen.clickSearchButton();
-    await SearchScreen.searchAndClickEmailBySubject(emailSubject);
-    await EmailScreen.checkOpenedEmail(senderName, emailSubject, emailText);
-    await EmailScreen.checkAttachment(encryptedAttachmentName);
-
-    await driver.terminateApp(bundleId);
-    await driver.activateApp(bundleId);
-
-    await MailFolderScreen.checkInboxScreen();
-    await MailFolderScreen.clickSearchButton();
-    await SearchScreen.searchAndClickEmailBySubject(emailSubject);
-
-    //try to see encrypted message with wrong pass phrase
-    await EmailScreen.enterPassPhrase(wrongPassPhrase);
-    await EmailScreen.clickOkButton();
-    await EmailScreen.checkWrongPassPhraseErrorMessage();
-
-    //check attachment after setting correct pass phrase
-    await EmailScreen.enterPassPhrase(correctPassPhrase);
-    await EmailScreen.clickOkButton();
-    await EmailScreen.checkOpenedEmail(senderName, emailSubject, emailText);
-    await EmailScreen.checkAttachment(encryptedAttachmentName);
-    await EmailScreen.clickOnAttachmentCell();
-    await AttachmentScreen.checkAttachment(attachmentName);
-
-    await AttachmentScreen.clickSaveButton();
-
-    await AttachmentScreen.checkDownloadPopUp(attachmentNameWithoutExtension);
-    await AttachmentScreen.clickCancelButton();
-    await AttachmentScreen.checkAttachment(attachmentName);
-    await AttachmentScreen.clickBackButton();
-
-    await EmailScreen.checkAttachment(attachmentName);
-    await EmailScreen.clickBackButton();
-
-    await MailFolderScreen.clickOnEmailBySubject(emailSubject);
-    await EmailScreen.checkOpenedEmail(senderName, emailSubject, emailText);
-    await EmailScreen.checkAttachment(encryptedAttachmentName);
-    await EmailScreen.clickOnAttachmentCell();
-
-    await AttachmentScreen.checkAttachment(attachmentName);
-
-    await AttachmentScreen.clickSaveButton();
-    await AttachmentScreen.checkDownloadPopUp(attachmentNameWithoutExtension);
-  });
-});
