+++ appium/tests/specs/mock/setup/FindEmailOnAttester.spec.ts
@@ -34,7 +34,7 @@ describe('SETUP: ', () => {
       await MailFolderScreen.checkInboxScreen();
       await MailFolderScreen.clickCreateEmail();
       await NewMessageScreen.setAddRecipient('available.on@attester.test');
-      await NewMessageScreen.checkAddedRecipientColor('Tom James Holub', 0, 'green');
+      await NewMessageScreen.checkAddedRecipientColor(CommonData.validMockUser.name, 0, 'green');
     });
   });
 });
