+++ appium/tests/specs/mock/setup/RespectsDisallowAttesterSearchForDomainsOnPerDomainBasic.spec.ts
@@ -43,7 +43,7 @@ describe('SETUP: ', () => {
 
       // Checking added recipients color
       await NewMessageScreen.checkAddedRecipientColor('attester@disabled.test', 0, 'gray');
-      await NewMessageScreen.checkAddedRecipientColor('Tom James Holub', 1, 'green');
+      await NewMessageScreen.checkAddedRecipientColor(CommonData.validMockUser.name, 1, 'green');
     });
   });
 });
